# ui/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Refactored free-fs with a subject-match-java style layered architecture.",
    "version": "v1.0.0"
  },
  "servers": [
    {
      "url": "http://192.168.164.18:30067"
    }
  ],
...
  "paths": {
...
    "/transfers/upload/init": {
      "post": {
        "summary": "初始化上传任务",
        "deprecated": false,
        "tags": [
          "FileTransferController"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InitTransferUploadCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SingleResultFileTransferTaskDTO"
                }
              }
            }
          }
        },
        "operationId": "initUpload",
        "parameters": []
      }
    },
    "/transfers/upload/check": {
      "post": {
        "summary": "上传前 MD5 校验",
        "deprecated": false,
        "tags": [
          "FileTransferController"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckUploadCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SingleResultCheckUploadResultDTO"
                }
              }
            }
          }
        },
        "operationId": "checkUpload",
        "parameters": []
      }
    },
    "/transfers/upload/chunk": {
      "post": {
        "summary": "上传分片",
        "deprecated": false,
        "tags": [
          "FileTransferController"
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "format": "binary",
                    "description": "分片文件",
                    "type": "string"
                  }
                },
                "required": [
                  "file"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SingleResultVoid"
                }
              }
            }
          }
        },
        "operationId": "uploadChunk",
        "parameters": [
          {
            "schema": {
              "format": "string",
              "type": "string"
            },
            "in": "query",
            "description": "任务ID",
            "required": true,
            "name": "taskId"
          },
          {
            "schema": {
              "format": "int32",
              "type": "integer"
            },
            "in": "query",
            "description": "分片索引",
            "required": true,
            "example": 0,
            "name": "chunkIndex"
          },
          {
            "schema": {
              "format": "string",
              "type": "string"
            },
            "in": "query",
            "description": "分片MD5（可选）",
            "required": false,
            "name": "chunkMd5"
          }
        ]
      }
    },
    "/transfers/upload/merge": {
      "post": {
        "summary": "手动触发合并",
        "deprecated": false,
        "tags": [
          "FileTransferController"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MergeChunksCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SingleResultFileRecord"
                }
              }
            }
          }
        },
        "operationId": "merge",
        "parameters": []
      }
    },
    "/transfers/download/init": {
      "post": {
        "summary": "初始化下载任务",
        "deprecated": false,
        "tags": [
          "FileTransferController"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InitDownloadCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SingleResultInitDownloadResultDTO"
                }
              }
            }
          }
        },
        "operationId": "initDownload",
        "parameters": []
      }
    },
    "/transfers/download/chunk": {
      "get": {
        "summary": "下载分片",
        "deprecated": false,
        "tags": [
          "FileTransferController"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/defaultPrimitive"
                }
              }
            }
          }
        },
        "operationId": "downloadChunk",
        "parameters": [
          {
            "schema": {
              "format": "string",
              "type": "string"
            },
            "in": "query",
            "description": "No comments found.",
            "required": true,
            "name": "taskId"
          },
          {
            "schema": {
              "format": "int32",
              "type": "integer"
            },
            "in": "query",
            "description": "No comments found.\nValidation[Min(value=0)]",
            "required": true,
            "example": 0,
            "name": "chunkIndex"
          }
        ]
      }
    },
    "/transfers/pause/{taskId}": {
      "post": {
        "summary": "暂停传输任务",
        "deprecated": false,
        "tags": [
          "FileTransferController"
        ],
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SingleResultVoid"
                }
              }
            }
          }
        },
        "operationId": "pause",
        "parameters": [
          {
            "schema": {
              "format": "string",
              "type": "string"
            },
            "in": "path",
            "description": "No comments found.",
            "required": true,
            "name": "taskId"
          }
        ]
      }
    },
    "/transfers/resume/{taskId}": {
      "post": {
        "summary": "恢复传输任务",
        "deprecated": false,
        "tags": [
          "FileTransferController"
        ],
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SingleResultVoid"
                }
              }
            }
          }
        },
        "operationId": "resume",
        "parameters": [
          {
            "schema": {
              "format": "string",
              "type": "string"
            },
            "in": "path",
            "description": "No comments found.",
            "required": true,
            "name": "taskId"
          }
        ]
      }
    },
    "/transfers/{taskId}": {
      "get": {
        "summary": "查询传输任务详情",
        "deprecated": false,
        "tags": [
          "FileTransferController"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SingleResultFileTransferTaskDTO"
                }
              }
            }
          }
        },
        "operationId": "getFileTransferByTaskId",
        "parameters": [
          {
            "schema": {
              "format": "string",
              "type": "string"
            },
            "in": "path",
            "description": "任务ID",
            "required": true,
            "name": "taskId"
          }
        ]
      },
      "delete": {
        "summary": "取消传输任务",
        "deprecated": false,
        "tags": [
          "FileTransferController"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SingleResultVoid"
                }
              }
            }
          }
        },
        "operationId": "cancelFileTransfer",
        "parameters": [
          {
            "schema": {
              "format": "string",
              "type": "string"
            },
            "in": "path",
            "description": "任务ID",
            "required": true,
            "name": "taskId"
          }
        ]
      }
    },
    "/transfers": {
      "get": {
        "summary": "按用户查询传输任务列表",
        "deprecated": false,
        "tags": [
          "FileTransferController"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MultiResultFileTransferTaskDTO"
                }
              }
            }
          }
        },
        "operationId": "listFileTransfersByUser",
        "parameters": [
          {
            "schema": {
              "format": "int32",
              "type": "integer"
            },
            "in": "query",
            "description": "No comments found.",
            "required": false,
            "example": 0,
            "name": "statusType"
          }
        ]
      }
    },
    "/transfers/files": {
      "get": {
        "summary": "按用户查询传输任务列表",
        "deprecated": false,
        "tags": [
          "FileTransferController"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MultiResultFileTransferTaskDTO"
                }
              }
            }
          }
        },
        "operationId": "listFileTransfersByUser_1",
        "parameters": [
          {
            "schema": {
              "format": "int32",
              "type": "integer"
            },
            "in": "query",
            "description": "No comments found.",
            "required": false,
            "example": 0,
            "name": "statusType"
          }
        ]
      }
    },
    "/transfers/pages": {
      "get": {
        "summary": "分页查询已完成传输任务",
        "deprecated": false,
        "tags": [
          "FileTransferController"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageResultFileTransferTaskDTO"
                }
              }
            }
          }
        },
        "operationId": "pageFileTransfersByUser",
        "parameters": [
          {
            "schema": {
              "format": "int32",
              "type": "integer"
            },
            "in": "query",
            "description": "No comments found.",
            "required": true,
            "example": 3,
            "name": "statusType"
          },
          {
            "schema": {
              "format": "string",
              "type": "string"
            },
            "in": "query",
            "description": "No comments found.",
            "required": false,
            "name": "keyword"
          },
          {
            "schema": {
              "format": "int64",
              "type": "number"
            },
            "in": "query",
            "description": "No comments found.",
            "required": true,
            "example": 0,
            "name": "pageIndex"
          },
          {
            "schema": {
              "format": "int64",
              "type": "number"
            },
            "in": "query",
            "description": "No comments found.",
            "required": true,
            "example": 20,
            "name": "pageSize"
          }
        ]
      }
    },
...
    "/transfers/chunks/{taskId}": {
      "get": {
        "summary": "查询已上传分片",
        "deprecated": false,
        "tags": [
          "FileTransferController"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MultiResultInteger"
                }
              }
            }
          }
        },
        "operationId": "getUploadedChunks",
        "parameters": [
          {
            "schema": {
              "format": "string",
              "type": "string"
            },
            "in": "path",
            "description": "No comments found.",
            "required": true,
            "name": "taskId"
          }
        ]
      }
    },
    "/transfers/download/chunks/{taskId}": {
      "get": {
        "summary": "查询已下载分片",
        "deprecated": false,
        "tags": [
          "FileTransferController"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MultiResultInteger"
                }
              }
            }
          }
        },
        "operationId": "getDownloadedChunks",
        "parameters": [
          {
            "schema": {
              "format": "string",
              "type": "string"
            },
            "in": "path",
            "description": "No comments found.",
            "required": true,
            "name": "taskId"
          }
        ]
      }
    },
...
    "/files": {
      "get": {
        "summary": "分页查询文件",
        "deprecated": false,
        "tags": [
          "FileController"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageResultFileRecordDTO"
                }
              }
            }
          }
        },
        "operationId": "search",
        "parameters": [
          {
            "schema": {
              "format": "string",
              "type": "string"
            },
            "in": "query",
            "description": "父目录ID",
            "required": false,
            "name": "parentId"
          },
          {
            "schema": {
              "format": "string",
              "type": "string"
            },
            "in": "query",
            "description": "搜索关键词",
            "required": false,
            "name": "keyword"
          },
          {
            "schema": {
              "format": "string",
              "type": "string"
            },
            "in": "query",
            "description": "No comments found.",
            "required": false,
            "name": "fileType"
          },
          {
            "schema": {
              "format": "boolean",
              "type": "boolean"
            },
            "in": "query",
            "description": "No comments found.",
            "required": false,
            "example": true,
            "name": "isDir"
          },
          {
            "schema": {
              "format": "boolean",
              "type": "boolean"
            },
            "in": "query",
            "description": "No comments found.",
            "required": false,
            "example": true,
            "name": "isRecents"
          },
          {
            "schema": {
              "format": "boolean",
              "type": "boolean"
            },
            "in": "query",
            "description": "是否已删除",
            "required": false,
            "example": true,
            "name": "deleted"
          },
          {
            "schema": {
              "format": "boolean",
              "type": "boolean"
            },
            "in": "query",
            "description": "是否收藏",
            "required": false,
            "example": true,
            "name": "favorite"
          },
          {
            "schema": {
              "format": "string",
              "type": "string"
            },
            "in": "query",
            "description": "排序字段",
            "required": false,
            "name": "sortField"
          },
          {
            "schema": {
              "format": "string",
              "type": "string"
            },
            "in": "query",
            "description": "排序方向",
            "required": false,
            "name": "sortOrder"
          },
          {
            "schema": {
              "format": "int64",
              "type": "number"
            },
            "in": "query",
            "description": "页索引",
            "required": true,
            "example": 0,
            "name": "pageIndex"
          },
          {
            "schema": {
              "format": "int64",
              "type": "number"
            },
            "in": "query",
            "description": "每页大小",
            "required": true,
            "example": 20,
            "name": "pageSize"
          }
        ]
      }
    },
    "/files/{fileId}": {
      "get": {
        "summary": "查询文件详情",
        "deprecated": false,
        "tags": [
          "FileController"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SingleResultFileRecordDTO"
                }
              }
            }
          }
        },
        "operationId": "getFileById",
        "parameters": [
          {
            "schema": {
              "format": "string",
              "type": "string"
            },
            "in": "path",
            "description": "文件ID",
            "required": true,
            "name": "fileId"
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "SingleResultFileTransferTaskDTO": {
        "type": "object",
        "properties": {
          "success": {
            "format": "boolean",
            "description": "状态",
            "type": "boolean",
            "example": true
          },
          "dataType": {
            "format": "string",
            "description": "返回数据格式",
            "type": "string"
          },
          "code": {
            "description": "返回状态码(object)",
            "type": "object"
          },
          "message": {
            "format": "string",
            "description": "返回消息",
            "type": "string"
          },
          "tips": {
            "format": "string",
            "description": "返回提示消息",
            "type": "string"
          },
          "data": {
            "description": "返回的单条数据(object)",
            "$ref": "#/components/schemas/FileTransferTaskDTO"
          }
        }
      },
...
      "CheckUploadCommand": {
        "type": "object",
        "properties": {
          "taskId": {
            "format": "string",
            "description": "No comments found.",
            "type": "string"
          },
          "fileMd5": {
            "format": "string",
            "description": "No comments found.",
            "type": "string"
          }
        },
        "required": [
          "taskId",
          "fileMd5"
        ]
      },
...
      "SingleResultCheckUploadResultDTO": {
        "type": "object",
        "properties": {
          "success": {
            "format": "boolean",
            "description": "状态",
            "type": "boolean",
            "example": true
          },
          "dataType": {
            "format": "string",
            "description": "返回数据格式",
            "type": "string"
          },
          "code": {
            "description": "返回状态码(object)",
            "type": "object"
          },
          "message": {
            "format": "string",
            "description": "返回消息",
            "type": "string"
          },
          "tips": {
            "format": "string",
            "description": "返回提示消息",
            "type": "string"
          },
          "data": {
            "description": "返回的单条数据(object)",
            "$ref": "#/components/schemas/CheckUploadResultDTO"
          }
        }
      },
...
      "defaultPrimitive": {
        "type": "object",
        "properties": {
          "shareIds": {
            "description": "分享ID列表,[array of string]",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "MergeChunksCommand": {
        "type": "object",
        "properties": {
          "taskId": {
            "format": "string",
            "description": "No comments found.",
            "type": "string"
          }
        },
        "required": [
          "taskId"
        ]
      },
...
      "SingleResultVoid": {
        "type": "object",
        "properties": {
          "success": {
            "format": "boolean",
            "description": "状态",
            "type": "boolean",
            "example": true
          },
          "dataType": {
            "format": "string",
            "description": "返回数据格式",
            "type": "string"
          },
          "code": {
            "description": "返回状态码(object)",
            "type": "object"
          },
          "message": {
            "format": "string",
            "description": "返回消息",
            "type": "string"
          },
          "tips": {
            "format": "string",
            "description": "返回提示消息",
            "type": "string"
          },
          "data": {
            "description": "返回的单条数据(object)",
            "type": "object"
          }
        }
      },
...
      "InitTransferUploadCommand": {
        "type": "object",
        "properties": {
          "userId": {
            "format": "string",
            "description": "No comments found.",
            "type": "string"
          },
          "fileName": {
            "format": "string",
            "description": "No comments found.",
            "type": "string"
          },
          "fileSize": {
            "format": "int64",
            "description": "No comments found.",
            "type": "number",
            "example": 0
          },
          "parentId": {
            "format": "string",
            "description": "No comments found.",
            "type": "string"
          },
          "totalChunks": {
            "format": "int32",
            "description": "No comments found.",
            "type": "integer",
            "example": 0
          },
          "chunkSize": {
            "format": "int64",
            "description": "No comments found.",
            "type": "number",
            "example": 0
          },
          "mimeType": {
            "format": "string",
            "description": "No comments found.",
            "type": "string"
          },
          "storageSettingId": {
            "format": "string",
            "description": "No comments found.",
            "type": "string"
          },
          "overwriteExisting": {
            "format": "boolean",
            "description": "No comments found.",
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "fileName",
          "fileSize",
          "totalChunks",
          "chunkSize",
          "mimeType"
        ]
      },
...
    }
  }
}